app.js restructuring, readability, Launchers added
Organized the code better, added 'type' property to both constructors to specify what equipment type they belong to, updated the function and method names for uniformity. 
Mostly fleshing out what the 'standard' for this project is to be shared by all equipment. There's a fair amount of redundancy to be overcome yet and the workload between class methods and functions isn't as clearly defined as I would like