Server Update and Equipment Lists
installed express files into the directory to prepare for putting the files
on to the raspberry pi, im completely unfamiliar with node and express
so this was mostly exploratory in nature.
I also appended the equipment list so each piece of equipment is now tracked
as well as adding a simple text file and trying to write and read from it.